Regression tests for github issue #234
In particular, terms printed into theory.sml files are done so with
the print_depth variable controlling how this happens.  If the user
has a print_depth variable that is set so as to cause a term to be
printed with '...'s occurring, then the term will fail to parse back
when the theory is loaded.