add manual pattern match heuristic / improve exhaustive one
Implement a new pattern match heuristic that allows manually
specifying in which order to do case splits. This one prints
a lot of debugging information in order to allow the user to
make decent choices. The techniques for producing this debug
have also been added to the exhaustive tactic. It now prints
debugging information on how far the exhaustive search
progressed and on intermediate results