deep_matches add multi-row redundancy detection
implement a check for redundant rows that considers multiple rows.
The previous one considers only single other rows and is therefore
faster, but not as powerful. Detecting redundancy is complicated and
for doing it some generation of nchotomy theorems using pattern
compilation was implemented. These nchotomies are handy for other
problems as well.