New theory of error-state monad.
Modelled on state_transformer example. Not committed to, or sure about
name of theory.  Didn't bring across the "monadic while" constant, but
prove most of the rest of things.  Also show it to be "applicative" as
per option, and the zipper example.

If  α → (β × α) option is "error state", is α → (β option × α) "state
error"?

Should probably write a functor or something to automate these to some
extent.