New example within examples/parsing: s-expressions
The s-expression type has constructors, SX_CONS, SX_STR, SX_NUM and
SX_SYM. There is pretty HOL-level syntax for the first of these, using

    ⦇ ... ⦈

to delimit lists, and with elements separated by ; as is usual for HOL
list-forms. (Such lists are internally terminated with SX_SYM "nil").
There is also notational support so that SX_CONS s1 s2 can be written

    ⦇ s1 • s2 ⦈

The SX_NUM constructor is admitted as a numeral form, so ``SX_NUM 10``,
can just be written as ``10``.

The PEG and grammar define what I think is a fairly traditional
Lisp-syntax, including 's as syntax for (quote s).  This latter is there
mainly for the hell of it, as I don't suppose we are likely to want to
give the type an operational semantics like Lisp's.

Getting the PEG to parse efficiently with EVAL is demonstrated and
tested in the self-test.  Some new generic facilities are moved into
pegLib and the pegScript.

Rather more is still possible.  For example, the provided PEG parses
directly from strings to sexps. In order to show a connection with the
sexp grammar in simpleSexpTheory, one would have to write another PEG
that parsed to parse-trees. One could then write functions to convert
from those parse-trees to sexps, and show that the composition of these
with the parse-tree parser gave the same result as the direct-to-sexp
parser.