Make new_definition_hook replace a constant on the LHS by a variable
new_definition calls gen_prim_specification which requires an equation
whose left-hand side is a variable, but sometimes (maybe only in obscure
old example script files) users pass an equation with a constant on the
left. The same hook that abstracts arguments now turns a constant on the
left into a variable.