Omit (deliberately) bogus set-logic command.
This causes Z3 4.4.0 to emit annoying warning message, and can't have
any useful semantics. Yices v1 doesn't object if this is left out.

Perhaps other solvers require a (set-logic) command, even if the
specified logic doesn't exist. If so, we will have to pass parameters to
selectively turn this text on or off.