Add CONTRIBUTORS list; modify COPYRIGHT
I've added a CONTRIBUTORS list that I hope is fairly comprehensive.
I've derived it from the list of acknowledgements in the Manual, and
from an inspection of our git history.

I've also modified the COPYRIGHT to simply assert ownership by the
CONTRIBUTORS.  As I've commented in the past, it may be simpler to try
to assert that it is all simply in the public domain.