Fix (somewhat obscure) looping bug in simplifier.
If the terms T = F and F = T both arrive in your assumption list, then
you will go into an infinite loop.  Neither should fire as a rewrite at
all.  Indeed, anything of the form  T = x, or F = x, is better off
flipped, unless it's the bad case above. In the bad case, treat the
rewrite as we treat a bare F.