update logging-kernel for gen_prim_specification and holy-hammer
Logging.sml is not updated, though (it will break in the defineConst
part).

Since HOL reduces defineConst to defineConstList outside of its kernel,
we will no longer target the former article command. However, it would
be possible to keep track of the fact that certain calls came from the
specialised version of new_definition, and give them the (slightly
shorter) proof... a task for another time.