work towards a compatibility theory with the OpenTheory stdlib
Currently this is a tedious manual process, but I'm not sure the best
way to proceed for automation.

OpenTheoryBoolTheory contains all the theorems in the OT stdlib bool
package (with no proofs).

HOL4CompatibilityTheory contains various theorems used in the HOL4
stdlib, proved, manually, in terms of the exports of
OpenTheoryBoolTheory.