Make monadsyntax more flexible with monad_unitbind
Previously, one had to bind "monad_unitbind" to a real constant, such as
option$OPTION_IGNORE_BIND. Now one can bind to a pattern, such as

  \m1 m2. OPTION_BIND m1 (K m2)

This means that one doesn't have to define as many constants when
working with a fresh monad, which feels like an advantage.

Also clean-up monad selftest, and test new behaviour.