Fix proof of "bigrec" thms for polymorphic records
These are the theorems that are proved to knit together the view of the
record that the user expects and the underlying reality, which is a
record of records.  These are stated and proved after the various
underlying record types are defined.

Tests in src/datatype/selftest.sml

This is work on github issue #173