Make local parsing work better for OmegaMath
Unfortunately, if you use Q.EXISTS_TAC and similar, you are dependent on
the global grammar (hidden behind a reference), and rebinding the Parse
structure doesn't protect you.  If this happens in a library, you really
have to use temp_set_grammars to get the right behaviour (and the
correct independence from what has happened globally).