Buffer logging of definitions
This is important when there is an .otd file indicating that certain
theorems should have their proofs deleted, and then certain
(non-skipped) definitions depend on those theorems. Definitions need to
be logged first, to ensure the article does not incorrectly refer to
external constants when it should be referring to defined constants.
But at definition time it is not yet known that a theorem is going to
have its proof deleted; previously, the proof-to-be-deleted would be
incorrectly logged when logging the definition that depends on it.

Now, definitions are buffered and only logged when export_thm is called,
so clients simply have to ensure that all proofs are deleted correctly
before calling export_thm (which is what they previously had to do
anyway).