Allow Hol_defn to accept more definitions.
It now accepts equations of the following form:

 test = \x. case x of [] => [] | _:: t => test t
 test f = \x. case x of [] => [] | _ :: xs => test f xs

where it would previous give a "failure in internal translation to tupled format" error.

The resulting definitions are

 ∀x. test x = case x of [] => [] | v::t => test t
 ∀x f. test f x = case x of [] => [] | v::xs => test f xs

As such, the user doesn't get back the requested equations in their original form. However, this may be preferable to failing completely.

(Some code has been tidied up, so this change looks larger than it actually is.)