Extend pp class of Const (vs SymConst) names
In particular, Const names get to be those consisting of alphanumerics +
apostrophes (primes) and underscores.  Previously, the presence of an
underscore or apostrophe would have pushed the name into the SymConst
class.

Treatment of these classes depends on the backend (and may be the same
for both) it's good to make the classes reasonable.