Improve logging of listTheory, axiomatise construction
Preivously, the construction via ind_type was leaking out via an
induction theorem in TypeBase that was never saved. The fix is to save
it (list_INDUCT0) and then use the otd file to delete its proof.

It would probably be possible to delete the binding at some point in
export_theory, to prevent it actually occurring in the final exported
theory, but that may not be worth the trouble.