Add targets for: OT base thms, unsat. HOL4 assums
The aim is to create an environment for proving the HOL4 base
assumptions using only the OT base theorems. The plan is to read the OT
theorems from an article as axioms, then read the HOL4 base assums from
an article as goals, and then write and record their proofs.