Update to new FunctionalRecordUpdate tech
This is from the mlton.org (as per URL in checked-in file). Also move
the source file to the Holmake directory so that I can use it there. The
Holmakefile in src/portableML then copies it from that location into
position so that it properly ends up in sigobj etc.

This version of the file supposedly compiles faster, and it also seems
to handle polymorphic fields.