Use state update functions to improve shell-cmds
The shell_commands function is really just a demonstrator for what I
want to do with Holmake, and this change makes it a bit more realistic.
In particular, there is now a maintained state on the client side
recording what's running, what's pending and what's completed. This gets
updated as the jobs complete, and the thinking is that Holmake would
maintain an analogous graph, enabling it to generate new jobs according
to the state of the graph.