Prove more hol4-base assumptions
There are now 13 left, and they all appear to require changes to the
setup. Some are to do with constants that are more specified in HOL than
OT (such as PRE and arithmetic$-). The rest are mostly to do with
boolTheory, which has not been recorded, and includes things like
TYPE_DEFINITION.