Make dep-graphs store commands as lists of strings
This will make it easier to extract and act on leading indicators such
as @ and - on each line.  If a line is

  -command

then it could be passed to the shell as command ; true.  If the command
has the leading @, then that line shouldn't be echoed to the shell.  In
a parallelised context, I'm not sure that anything much should be echoed
to the user anyway.

Previously, I had Holmake concatenate the lines from the Holmakefile
with intervening semicolons, which would just make extracting the @ and
- characters that much harder.