Reduce parse_print to three cheats
The main one is about being able to parse one sexp followed by a
separating character (dot or space or close-paren) and have this be the
same as just parsing the sexp on its own (leaving the rest of the string
unconsumed).

The other cheats are about using that lemma for parsing some
space-separated lists of sexps.