Emit CHEAT when non-qof tactic fails; get nice colours
By having the tactic cheat, and the non-qof handling of prove both emit
the string "CHEAT" as part of their behaviour, Holmake's multibuild can
give you a CHEATED verdict on scripts that do either.  Strictly,
speaking the tag attached to non-qof failure is not cheat, but I don't
think that's really important here.