Rework Poly/ML quotation use to do in-process lexing
This changes from having to use a separate unquote process as part of
Holmake, and may help resolve github issue #340

The drawback at the moment is that the whole file is sucked into memory
before Poly/ML gets to see it. The way the filter code works
could (should?) be changed to lex things in a demand-driven way.