Provide CLINE_OPTIONS variable in Holmakefiles
This variable is read for a set of default command-line switches which
are treated as if they are being passed on the command-line. Things
actually passed on the command-line are applied after what is in the
Holmakefile, possibly overriding the file's switches. Of course, the
command-line can also change the defaults with both of the
--no_hmakefile and --holmakefile=file switches.

Motivated by the desire to have Holmakefiles be able to contain good -j
options for specific directories.

This pretty well supercedes the OPTIONS variable, so the documentation
in DESCRIPTION has been adjusted accordingly.

Closes #86