Handle opentheory intermediate leftovers better
In particular, the parallel process for building .art files creates an
intermediate thyname.artScript.sml file that shouldn't be regarded as a
normal script file.  (Indeed, it can't work as a normal script file
because of the presence of the full-stop in the theory name, so there's
a warning emitted for other occurrences of this if seen.)

These intermediates can be cleaned by Holmake as well if they're
found.  (Normal execution of Holmake will remove these files after
they're used, but if the build is interrupted for some reason, they get
left behind.)