Strengthen statement of TAKE_SUM
Also tidy up listScript, removing a big local-in-end that made bit
cut-and-pastes painful. (If you put your point down in the middle of a
local-in-end, and do a cut-and-paste, you will get an syntax error.) For
this many theorems at once, better just to export the local declarations
to the top-level and deal with the consequent changes.

One theorem needed to have its proof adjusted.

Closes #330.