Patch glitch in Holmake output.
The "usepfx" flag is no longer set when `CLINE_OPTIONS = -j1` appears in the Holmakefile. It will only be set when `-j1` is in the actual command line.

This avoids lines that is too long (due to `Holmake: ` being printed) when recursing into a directory that doesn't have  `CLINE_OPTIONS = -j1` in it's Holmakefile.

I suspect `Holmake: ` is not printing when recursing into a directory that does have `CLINE_OPTIONS = -j1`, but this wouldn't result in lines that are too long.