Move patternMatchesExample to .ML suffix
As it stands this file is mal-formed for compilation with Moscow ML (and
it is never compiled with Poly/ML because it is never referenced
anywhere). As the intention is have this file be something a user plays
with interactively, we don't want Holmake to look at it at all.  The way
to achieve this is to use the .ML suffix.