Fix grammarLib in light of changes to errormonad
This code relied on the preservation of state even in the presence of
failure so that it could store information about where failures had
occurred during parsing.

Rather than try to figure out how to do this with the revised
errormonad (which doesn't preserve failing states), just redefine the
monad in the prelude to grammarLib so that the rest of the code works.