Fix simpleSexpParse by removing unused & failing proof
Because the theorem wasn't saved, or used anywhere in the script,
Holmake didn't see a CHEAT-line and so didn't print an orange CHEATED
message or similar. Nor was the exported theory file any guide that a
problem had arisen. Nonetheless, when run with --qof, as done by build,
the script failed. The CHEAT line does appear in the .hollog, and this
may be an argument for better filtering by the (Poly/ML) monitor
process.