Fix wordsScript in light of pat_assum rename
One interesting case requires a more specific pattern than earlier
because a purely general pattern has the parser guessing the wrong
addition-constant when the pattern looks to general (and this happens
before any matching is attempted).

I want to deal with this situation by having the parse-in-context
primitives try all possible parses when generating patterns.