Fix bug in TFL on experimental kernel
Problem is ultimately that the strip_forall/SPEC_ALL pair match up on
the standard kernel (both rename variables with same names and different
types), but they don't match up on the experimental kernel as
strip_forall doesn't rename, and SPEC_ALL does.

Change a local use of SPEC_ALL (in the code that fixes up nested
recursion inductions) to be a non-renaming version and fix the behaviour
on the experimental kernel.

Include a test-case.