Display stalling info in one-line Holmake statuses
If there is text scrolling past, then it's clear that something is
happening (and what it is), but if the process stalls, one only used to
see the last line of text.  Now, that line of text will pick up the
stall-indicators (up to 3 red exclamation marks) if the underlying
process does nothing for a while.