Make more functions in Q use Q_TAC and QTY_TAC
This should make the behaviours more uniform, and certainly makes the
definitions shorter.

By using Q_TAC, code can backtrack over multiple possible parses (which
are caused by overloading). In places where success/failure can't be
distinguished by subsequent backtracking, the old code is retained. For
example, in Q.ABS, once you have a successful parse of the right type,
the use of the parsed term in SPEC is always going to succeed.

The exception is the definition of ISPECL_THEN, which stays complicated
because you have to parse the list of terms all at once, and pass them
to ISPECL. As per the comment in Drule, just iterating ISPEC "DOESN'T
WORK". (This is because you want to calculate one type-instantiation all
at once. Iterating may cause later type variables to get identified with
earlier ones by earlier parts of the instantiation.)