Really stop other jobs in absence of -k flag
If Holmake had more runnable jobs than it could run (so they weren't
currently running because the parallelism limit had been reached), and a
running job died with an error, the jobs currently running *would* all
get killed, but the other, pending jobs would then get pulled in for
execution.

Without the -k flag present, this is incorrect.

Fix this behaviour by changing the worklist structure's genjob function
to be a constant NoMoreJobs value when it receives the KillAll command
from the client.