Refactor handling of ancestral grammars
The eventual aim is to make it easy for users to override the selection
of ancestral grammars. At the moment, you are forced to have the union
of the grammars from all the ancestral theories. This gets unreasonable
in the presence of heaps, which may include a whole bunch of theories
that you don't want to have messing up your parsing context.