Remove "C" from combinTheory.combin_grammars
Currently, combinTheory is set up to hide C from the global grammar when
it is loaded, but to leave C in combinTheory's own grammar, which you
get via writing things like combinTheory.combin_grammars. This is weird.
To make things more consistent, pull C out of combin's grammar
properly (without having to use adjoin_to_sml), and have it absent in
both descendants, and in combin_grammars.

Of course, the standard trick of typing

  reveal "C"

makes it visible again.

The code in src/metis/normalForms.sml used this visibility when stating
and proving a rewrite about C. This was fixed simply by making the
statement use the combin$C notation.