Fix a thm being stored under the wrong name
univ_reln_to_rel_conv was being stored as "rrestrict_rel_conv". Not only
did the stored name not match the ML level binding, it shadowed an
existing name, thereby causing the old theorem to not appear in the
Theory file at all.

Thanks to Thibault Gauthier for the bug report.