Fix a Holmakefile to dodge buildheap's new pickiness
Issue was that INCLUDES line quoted absolute paths for directories, but
the buildheap commandline referenced object files using relative paths.
The behaviour in buildheap could be made smarter, so that it spotted
that the paths were pointing to the same place, but I prefer to drop the
absolute paths and just use relative paths in the INCLUDES line.

Relative paths have two advantages:
1. developments can be moved around more readily
2. by avoiding HOLDIR in INCLUDES lines, possible problems if HOLDIR
   includes spaces are avoided (may not really be an issue).

Having the textual forms be the same might allow for factorisation
through a makefile variable.