Revert two recent changes to Defn.sml
It turns out some CakeML code depends on the internal workings of Defn,
including the fact that certain constants are defined with certain names
chosen. So, at least for the moment, it's not a good idea to have the
changes to Defn.sml from d680453 and fa09c98.

The machinery from those commits allowing the changes to Defn seems
reasonable to keep, even if it's not being used right now.