pattern_matches library: use new syntax everywhere
After Michael's work on how to parse PMATCH expressions nicely, they use
the same syntax as decision tree based case expressions. The old syntax
something like `CASE ... OF [...; ...; ...]` therefore needs updating to
`case ... of | ... | ... | ...`.

This commit changes the syntax of case expressions in the examples files
and comments. This did not cause problems before, because this code is
only run manually either during debugging or as a documentation aid.

Moreover, some code that is unused due to Michael's changes now is
removed by this commit.