patternMatchesLib: avoid replacing double bound vars by wildcards
The simplifications for PMATCH case expressions replaces variables only
used in the pattern with wildcards. If the variable is used multiple
times in the pattern, this leads to confusing output. The pretty-printed
version does not allow to identify these double binds any more (while
they are still present in the underlying term). This commit fixes this
issue.

For example, before ``case l of | x::x::_ => T | _ => F`` was simplified
to ``case l of | _::_::_ => T | _ => F``. This looks like any list with
at least 2 elements was excepted. The internal representation used the
same wildcard var for the first 2 elements though and was therefore
correct.