patternMatchesLib: fix bug with looping PMATCH simplification
If trying to discard precondition via "rc_ss" called the
PMATCH_SIMP_CONV itself, the procedure was looping for some inputs.
While this bug was in principle there before, this was now very
prominent due to adding the PMATCH_SIMP_ss to srw_ss with commit
be55e2c.

An appropriate test-case will be added to the selftest in a later commit
together with other selftest changes.