decision-tree case expressions: fail when mixing literals and constructor patterns
Before, it was possible to mix literal and constructor patterns in one
column. The results were not as intended though. For example

``case x of 0 => 0 | 1 => 1 | SUC _ => 2``

is compiled to (with pretty printing turned off)

``literal_case
    (λv.
       if v = 0 then 0
       else if v = 1 then 1
       else if v = SUC m then 2
       else ARB) x``

The pattern "SUC m" is handled as a value. "m" is a free variable of the
resulting case expression. It is not - as most likely intended by the
user - matched against "x".

The problem is that this case expression mixes literals and constructor
patterns. The documentation states explicitly that this is an
unsupported case expression. I thought the parser would fail on such
case expressions. I suspect (but am uncertain and could not find
anything going through the old versions) that it did at some point.
Anyhow, this commits adds such a check during pattern compilation now
that prevents such patterns from being parsed.