Rename file in sep. logic example
Having a file called header.sml confused dependency calculation on Macs
because its calculation of what happened in the .grm.sml file thought
there was a dependency on a structure called Header that was actually
purely internal. Because of the Mac's case-insensitive file system, this
was thought to be a dependency on the header file.  This led to build
failure.