Print exceptions produced by QUse.use in Vimhol
My understanding of what's going on is that the exception is raised
inside the runner thread correctly, but is discarded as the thread dies.
Therefore, we now explicitly print it first before dying.

Maybe QUse.use should be printing exceptions generally (like use does).
But that would be a bit ugly. The fact that use prints exceptions is why
explicitly printing in Vimhol was not previously needed.

One minor issue with this change is that if one sends a call to "use" via
Vimhol, and the used file raises an exception, it will be printed twice.
That seems better than not printing anything in the common case, though.