Ensure PolyML interactive sessions use correct "use"
There were two implementations hanging around, and only one (the one
in QUse) correctly used "unquoting" when the file being used contained
Unicode but no back-tick characters.

Delete all of the bad implementation and bind use to QUse.use in the
prelude.ML file to make sure that this use is the right one. This fix
makes 1ec1cd7bf7c unnecessary (I hadn't appreciated that this was the
problem Ramana was encountering).