Reorganise flow of overloading resolution
This should let more specific errors escape the process. Now, the test
case in the selftest lets a (located) error about the impossible
interaction between < and T escape to the user. Previously, 3 + 4 <
T (with < and + overloaded to integer and number possibilities) was
just generating the "no resolution of overloads possible" message
without a helpful location.