Stop Holmake generating default targets from *Theory.{sig,sml}
If a directory has a fooScript.sml file, then fooTheory.uo will still
be generated as a default target, so there's no need to generate
fooTheory.uo from fooTheory.sml. If fooTheory.sml is stale or
orphaned (perhaps its script file has been moved to another
directory), attempting to build in this directory will just cause an
annoying Holmake error. If fooScript.sml is generated from a special
rule in a Holmakefile, then that Holmakefile can equally take
responsibility for making fooTheory.uo a target.

This is part of an approach to github issue #295.

The second part of this would be automatically remove the
stale/orphaned theory files.