Let hol.state0 be rebuilt with an older time-stamp
This mimicks what was done in the buildheap utility for all of HOL's
other heaps (in df4694a). This is progress towards github issue #105.
Now heaps can be deleted without needing *Theory.sml files to be
rebuilt. Unfortunately, a HOL tree can't be relocated until .uo and
.ui files are adjusted because these have absolute filenames in them.

So far this work has all been for Poly/ML (Moscow ML doesn't have
heaps, and its .uo files are somewhat location-independent, as long as
the loadPath is set up so that the loader/linker can find them).