Rebuild .uo and .ui files without forcing unnecessary extra work
This is by analogy with what has already been attempted for heaps, and
is progress with github issue #105.

Also fix a bug with the heap handling if the heap was in a
directory with script-files (not the case with the heaps in HOL/src):
the dependency graph built for the directory spotted that there was a
dependency from the heap to the script files, and even though the
heap's time-stamp was set appropriately, the existence of the link in
the graph meant that the rest of the graph got executed.

The fix for this (in both the case of the heap and the .uo/.ui file)
is for execution of graph nodes to check at the time the node is made
ready to execute if the target really does need to be redone (by
examining time-stamps). Then, the rebuild can be skipped even though
the initial analysis thought it was going to have to happen.

Future work:
- For the moment, both build and Holmake rely on having a particular
  environment variable set to enable this option. I will augment
  and/or replace this with command-line options. (The buildheap
  program needs to retain this use of the environment variable though
  because neither Holmake nor build control how that program is
  called; instead this program is called by the user via
  Holmakefiles.)
- Build needs to take an option to clean a HOL installation of its
  heaps and ui/uo files. I don't know how to automatically identify
  the heaps, but could just purge those that are identified as such in
  Holmakefiles.