emacs mode: provide support for de-Unicode-ify
There is a new command `hol-remove-unicode` now that replaces commonly
used unicode symbols with their ASCII counterpart. It can be found in
the menu as "HOL / Misc / Replace common HOL unicode symbols". If a
region is selected, it operates only on this region, otherwise on the
whole buffer.

Currently the most common unicode symbols are replaced. Probably a few
more need adding in time. This should be straightforward by updating the
list "hol-unicode-replacements".