Allow users to have theorems called "print"
Closes #403

The solution works for print, but hasn't been generalised to cover all
possible problems of this form yet. A brief inspection suggests that
'print' *may* have been the only name not appropriately
protected. (Approach is to make sure that we use qualified names in ML
that appears after the user's theorems have been declared.)

Test-case included.